refactor(dev): make `group` and `endGroup` method reusable to other command classes
